requirements.txt
requests
orm[aiosqlite]
aiosqlite
python-bale-bot